button removed and worked on payment.js
Right now the home page is edited for better ui.

Payment.js is created.

is saying to create a function to support it but im not too sure on that.

at firebase.js i have also changed so it can get a payment intent id from stripe.

publishable key (from app.js) is stored in .env